feat(GN-71): 프로필 이미지 업로드 후 저장 구현
